Factorize code of user impersonation
